refactor: move out tree_sitter specific stuff
