+++ crates/config/src/combined.rs
@@ -266,7 +266,7 @@ mod test {
   use crate::from_str;
   use crate::test::TypeScript;
   use crate::SerializableRuleConfig;
-  use ast_grep_core::language::LanguageExt;
+  use ast_grep_core::LanguageExt;
 
   fn create_rule() -> RuleConfig<TypeScript> {
     let rule: SerializableRuleConfig<TypeScript> = from_str(
