+++ crates/config/src/transform/transformation.rs
@@ -207,7 +207,7 @@ mod test {
   use super::*;
   use crate::test::TypeScript;
   use crate::DeserializeEnv;
-  use ast_grep_core::language::LanguageExt;
+  use ast_grep_core::LanguageExt;
   use serde_yaml::with::singleton_map_recursive;
   use std::collections::HashMap;
