+++ xtask/src/schema.rs
@@ -1,7 +1,8 @@
 use anyhow::{bail, Context, Result};
 use ast_grep_config::SerializableRuleConfig;
-use ast_grep_core::language::{Language, LanguageExt, TSLanguage};
+use ast_grep_core::language::TSLanguage;
 use ast_grep_core::matcher::{Pattern, PatternBuilder, PatternError};
+use ast_grep_core::{Language, LanguageExt};
 use ast_grep_language::{
   Alias, Bash, CSharp, Cpp, Css, Elixir, Go, Haskell, Html, Java, JavaScript, Json, Kotlin, Lua,
   Php, Python, Ruby, Rust, Scala, Swift, Tsx, TypeScript, Yaml, C,
